[tset_ld.c] maybe this will fix the issue for i686-freebsd? 	    Cf http://hydra.nixos.org/build/5662697.
[From SVN r8658 (trunk)]

+++ tests/tset_ld.c
@@ -106,7 +106,7 @@ print_binary (long double d, int flag)
   /* now e <= d < 2e */
   if (flag)
     printf ("0.");
-  while (d > 0)
+  while (d > (long double) 0.0)
     {
       prec++;
       if (d >= e)
